bot/common: Improve updating sources
Previously you could only specify a remote from which
you could update the sources and you could only use 'master'
branch. Moreover 'reset --hard' had been used to do an update
so there was a possibility of losing some changes.

This patch introduces an improvement to the mechanism of
updating source code repositories. There is a new section
in the bot configuration that specifies what repositories are to
be updated. This functionality supports multiple repositories.
You can specify a path to the repository, absolute
or relative, a remote and a branch. You can also specify
if you want to stash non-committed changes. If the value
is False and there are changes then the repository will not be updated.