Revert "ptscontrol: Terminate Protocol Viewer process along with PTS"
This reverts commit e5ee02fce18a3ea994145411d30627ede5b69a62.

The protocol viewer allocates the buffer file in local disk every time
when it starts. But terminating the protocol viewer process doesn't
remove the buffer file. So, aftre running the full cycle, it fills up
the disk space with the left over buffer files and eventually the system
hangs with out of space.

Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>